# docs/swagger.yaml
definitions:
  models.Application:
    properties:
      createdAt:
        type: string
      fileURL:
        type: string
      id:
        type: integer
      status:
        type: string
      text:
        type: string
      updatedAt:
        type: string
      userID:
        type: integer
    type: object
  models.CreateApplicationRequest:
    properties:
      file_url:
        type: string
      status:
        type: string
      text:
        type: string
    required:
    - file_url
    - text
    type: object
  models.UpdateApplicationRequest:
    properties:
      file_url:
        type: string
      status:
        type: string
      text:
        type: string
    type: object
host: localhost:8081
info:
  contact: {}
  description: API приложения Application с JWT авторизацией
  title: Application service
  version: "1.0"
paths:
  /api/applications:
    get:
      consumes:
      - application/json
      parameters:
      - description: Filter by user ID
        in: query
        name: user_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Application'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Gets all the Applications
      tags:
      - UserApplication
    post:
      consumes:
      - application/json
      parameters:
      - description: Create Application
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.CreateApplicationRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Application successfully created
          schema:
            $ref: '#/definitions/models.Application'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create Application
      tags:
      - UserApplication
  /api/applications/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Application ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete Application by ID
      tags:
      - UserApplication
    get:
      consumes:
      - application/json
      parameters:
      - description: Application ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Данные одной заявки
          schema:
            $ref: '#/definitions/models.Application'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Gets Application by ID
      tags:
      - UserApplication
    patch:
      consumes:
      - application/json
      parameters:
      - description: Application ID
        in: path
        name: id
        required: true
        type: integer
      - description: Application request with user_id, text и status
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.UpdateApplicationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Application'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update Application
      tags:
      - UserApplication
schemes:
- http
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"